Fix decoder forward
