+++ train_rqvae.py
@@ -2,29 +2,14 @@ import gin
 
 from accelerate import Accelerator
 from data.movie_lens import MovieLensMovieData
-from data.schemas import SeqBatch
+from data.utils import cycle
+from data.utils import next_batch
 from distributions.gumbel import TemperatureScheduler
 from modules.rqvae import RqVae
 from torch.optim import AdamW
 from torch.optim.lr_scheduler import LinearLR
 from torch.utils.data import DataLoader
 from tqdm import tqdm
-
-
-def cycle(dataloader):
-    while True:
-        for data in dataloader:
-            yield data
-
-
-def next_batch(dataloader, device):
-    batch = next(dataloader)
-    return SeqBatch(
-        user_ids=batch.user_ids.to(device),
-        ids=batch.ids.to(device),
-        x=batch.x.to(device),
-        seq_mask=batch.seq_mask.to(device)
-    )
 
 
 @gin.configurable
